Lowered Cooldown to 3s
This way it's more compatible with glitchless, as there are multiple instances where the autosplitter may or may not split depending on how fast you are, now it will always split. This also removes one of the uses of the counters for the experimental setting, so the fast travel after Bane's hideout will always split. It also adds a couple of extra splits. (As a side note, the normal any% splits and experimental ones would now be identical except for the extended cooldown in the sewers).